Refactor classes to separate packages
