+++ src/main/java/commands/Command.java
@@ -1,3 +1,10 @@
+package commands;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import exceptions.DukeException;
+
 /**
  * Represents the commands that are available in the chatbot
  */
