+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,7 @@
+package duke;
+
+import task.Task;
+
 import java.util.ArrayList;
 import java.util.List;
