+++ src/main/java/task/Todo.java
@@ -1,3 +1,5 @@
+package task;
+
 /**
  * Represents a todo, a type of task
  */
@@ -8,7 +10,7 @@ public class Todo extends Task {
      * @param description The specified description.
      * @param isDone      The boolean indicating whether the task is done.
      */
-    Todo(String description, boolean isDone) {
+    public Todo(String description, boolean isDone) {
         super(description, isDone);
         this.taskType = TaskType.T;
     }
